# colblk: remove use of UnsafeRawSlice in UintBuilder