#### cockroachkvs: reduce number of benchmarks

Using `benchdiff` for this package is impractical, there are many
hundreds of config variations across the benchmarks. This commit
switches to using a small number of configs instead.

#### colblk: remove use of UnsafeRawSlice in UintBuilder

There is not much benefit to using an unsafe slice rather than a
regular slice. We still use an unchecked get (except that we can check
the bounds in invariant builds).

```
name                                                                                                                                         old time/op    new time/op    delta
CockroachDataColBlockWriter/AlphaLen=8,Prefix=8,Shared=4,KeysPerPrefix=4,Logical=10,ValueLen=8-24                                               139µs ± 1%     140µs ± 1%    ~     (p=0.853 n=10+10)
CockroachDataColBlockWriter/AlphaLen=8,Prefix=128,Shared=64,KeysPerPrefix=4,Logical=50,ValueLen=128-24                                         32.3µs ± 2%    32.4µs ± 2%    ~     (p=0.971 n=10+10)
CockroachDataColBlockWriter/AlphaLen=26,Prefix=1024,Shared=512,KeysPerPrefix=1,ValueLen=1024-24                                                5.59µs ± 9%    5.60µs ±10%    ~     (p=0.754 n=10+10)
```